+++ l3_footstep_planning_plugins/src/std/robot_model/support_polygon_check.cpp
@@ -34,7 +34,7 @@ bool SupportPolygonCheck::loadParams(const ParameterSet& params)
   polygons_.clear();
 
   // iterate through all feet
-  for (size_t i = 0; i < p.size(); i++)
+  for (int i = 0; i < p.size(); i++)
   {
     FootIndex foot_idx;
     if (!getYamlValue(p[i], "idx", foot_idx))
